A Restful Web Service in Java
This is a basic restful web service application that was created by
walking through the videos from Java Brains on youtube. This application
uses the concept of a social network where there are Profiles, Messages,
Comments. There are relevant objects and the Resource and Service
classes created to support the Restful Web Service.